voice-assistant/requirements.txt
faster-whisper>=1.0.0
piper-tts>=1.2.0
numpy>=1.24.0
scipy>=1.10.0
websockets>=12.0
requests>=2.31.0
